restructure into transformer.py
